+++ apps/backend/src/modules/user/modules/auth/auto-auth.controller.ts
@@ -8,29 +8,48 @@ import {
 	UseGuards,
 } from '@nestjs/common';
 import { QueryBus } from '@nestjs/cqrs';
+import {
+	ApiBearerAuth,
+	ApiInternalServerErrorResponse,
+	ApiOkResponse,
+	ApiOperation,
+	ApiTags,
+	ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
 
 import { Public } from '@/decorators/public.decorator';
 import { CurrentUserEmail } from '@/decorators/current-user-email.decorator';
 
 import { AuthService } from './auth.service';
-import type { IAutoLoginResponse } from './interfaces/responses';
+import { AutoLoginResponse } from './classes/responses';
 import { RefreshTokenGuard } from './guards/refresh-token.guard';
 import { JwtTokenType } from './models/jwt-token';
 import Routes from './auth.routes';
 import { AutoAuthContract } from './queries/contracts/auto-auth.contract';
 import type { IAutoAuthLoggedUser } from './interfaces/user';
 
+@ApiTags('Auth')
 @Controller(Routes.CONTROLLER)
 export class AutoAuthController {
 	private readonly logger = new Logger(AutoAuthController.name);
 
 	constructor(private readonly queryBus: QueryBus, private readonly authService: AuthService) {}
 
+	@ApiBearerAuth('refresh-token')
+	@ApiOperation({ description: 'Auto authentication if providing refresh token' })
+	@ApiOkResponse({
+		description: 'If successfully fulfill the auto auth request',
+		type: AutoLoginResponse,
+	})
+	@ApiUnauthorizedResponse({
+		description: 'If could not find a logged user, or refresh token is either missing or invalid',
+	})
+	@ApiInternalServerErrorResponse({ description: 'If failed to retreive logged user' })
 	@Public()
 	@UseGuards(RefreshTokenGuard)
 	@Get(Routes.AUTO_AUTH)
 	@HttpCode(HttpStatus.OK)
-	public async autoAuth(@CurrentUserEmail() userEmail: string): Promise<IAutoLoginResponse> {
+	public async autoAuth(@CurrentUserEmail() userEmail: string): Promise<AutoLoginResponse> {
 		this.logger.log(`Will try to auto auth with data email: "${userEmail}"`);
 
 		const loggedUser = await this.queryBus.execute<AutoAuthContract, IAutoAuthLoggedUser | null>(