+++ apps/backend/src/modules/user/modules/auth/github.controller.ts
@@ -2,6 +2,7 @@ import {
 	BadRequestException,
 	Controller,
 	Get,
+	HttpStatus,
 	Logger,
 	Redirect,
 	UseFilters,
@@ -10,6 +11,7 @@ import {
 import { CommandBus, EventBus, QueryBus } from '@nestjs/cqrs';
 import { RealIP } from 'nestjs-real-ip';
 import { ConfigService } from '@nestjs/config';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
 
 import { Public } from '@/decorators/public.decorator';
 import { ExternalAuthUser } from '@/decorators/external-auth-user.decorator';
@@ -29,6 +31,7 @@ import { LoginMixpanelContract } from './events/contracts/login-mixpanel.contrac
 import { JwtTokenType } from './models/jwt-token';
 import { ExternalAuthFilter } from './filters/external-auth.filter';
 
+@ApiTags('Auth')
 @Controller(Routes.CONTROLLER)
 export class GithubController {
 	private readonly logger = new Logger(GithubController.name);
@@ -41,6 +44,7 @@ export class GithubController {
 		private readonly configService: ConfigService<IEnvironment, true>,
 	) {}
 
+	@ApiOperation({ description: 'A redirect URL to enter GitHub OAuth app' })
 	@Public()
 	@UseGuards(GithubAuthGuard)
 	@Get(Routes.GITHUB_AUTH)
@@ -48,11 +52,14 @@ export class GithubController {
 		return;
 	}
 
+	@ApiOperation({
+		description: 'A redirect URL used by GitHub to send back the server the user data',
+	})
 	@Public()
 	@UseGuards(GithubAuthGuard)
 	@UseFilters(ExternalAuthFilter)
 	@Get(Routes.GITHUB_REDIRECT)
-	@Redirect(undefined, 301)
+	@Redirect(undefined, HttpStatus.MOVED_PERMANENTLY)
 	public async githubRedirect(@ExternalAuthUser() user: IExternalAuthUser, @RealIP() ip: string) {
 		this.logger.log(
 			`User with an email "${user.email}" tries to auth. Will check if already exists in DB`,