+++ apps/backend/src/modules/user/modules/auth/google.controller.ts
@@ -2,7 +2,6 @@ import {
 	BadRequestException,
 	Controller,
 	Get,
-	HttpCode,
 	HttpStatus,
 	Logger,
 	Redirect,
@@ -12,6 +11,7 @@ import {
 import { CommandBus, EventBus, QueryBus } from '@nestjs/cqrs';
 import { RealIP } from 'nestjs-real-ip';
 import { ConfigService } from '@nestjs/config';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
 
 import { Public } from '@/decorators/public.decorator';
 import { ExternalAuthUser } from '@/decorators/external-auth-user.decorator';
@@ -31,6 +31,7 @@ import { LoginMixpanelContract } from './events/contracts/login-mixpanel.contrac
 import { JwtTokenType } from './models/jwt-token';
 import { ExternalAuthFilter } from './filters/external-auth.filter';
 
+@ApiTags('Auth')
 @Controller(Routes.CONTROLLER)
 export class GoogleController {
 	private readonly logger = new Logger(GoogleController.name);
@@ -43,6 +44,7 @@ export class GoogleController {
 		private readonly configService: ConfigService<IEnvironment, true>,
 	) {}
 
+	@ApiOperation({ description: 'A redirect URL to enter Google OAuth app' })
 	@Public()
 	@UseGuards(GoogleAuthGuard)
 	@Get(Routes.GOOGLE_AUTH)
@@ -50,12 +52,14 @@ export class GoogleController {
 		return;
 	}
 
+	@ApiOperation({
+		description: 'A redirect URL used by Google to send back the server the user data',
+	})
 	@Public()
 	@UseGuards(GoogleAuthGuard)
 	@UseFilters(ExternalAuthFilter)
 	@Get(Routes.GOOGLE_REDIRECT)
-	@HttpCode(HttpStatus.OK)
-	@Redirect(undefined, 301)
+	@Redirect(undefined, HttpStatus.MOVED_PERMANENTLY)
 	public async googleRedirect(@ExternalAuthUser() user: IExternalAuthUser, @RealIP() ip: string) {
 		this.logger.log(
 			`User with an email "${user.email}" tries to auth. Will check if already exists in DB`,