+++ apps/backend/src/modules/user/modules/auth/interfaces/responses.ts
@@ -1,9 +1,0 @@
-import type { IAutoAuthLoggedUser } from './user';
-
-export interface IRefreshTokenResponse {
-	accessToken: string;
-}
-
-export interface IAutoLoginResponse extends IAutoAuthLoggedUser {
-	accessToken: string;
-}