+++ apps/backend/src/modules/user/modules/auth/refresh-token.controller.ts
@@ -1,4 +1,12 @@
 import { Controller, Get, HttpCode, HttpStatus, Logger, UseGuards } from '@nestjs/common';
+import {
+	ApiBearerAuth,
+	ApiInternalServerErrorResponse,
+	ApiOkResponse,
+	ApiOperation,
+	ApiTags,
+	ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
 
 import { Public } from '@/decorators/public.decorator';
 import { CurrentUserId } from '@/decorators/current-user-id.decorator';
@@ -7,15 +15,26 @@ import { CurrentUserEmail } from '@/decorators/current-user-email.decorator';
 import { RefreshTokenGuard } from './guards/refresh-token.guard';
 import { AuthService } from './auth.service';
 import { JwtTokenType } from './models/jwt-token';
-import type { IRefreshTokenResponse } from './interfaces/responses';
+import { RefreshTokenResponse } from './classes/responses';
 import Routes from './auth.routes';
 
+@ApiTags('Auth')
 @Controller(Routes.CONTROLLER)
 export class RefreshTokenController {
 	private readonly logger = new Logger(RefreshTokenController.name);
 
 	constructor(private readonly authService: AuthService) {}
 
+	@ApiBearerAuth('refresh-token')
+	@ApiOperation({ description: 'Get a new access token by providing a refresh token' })
+	@ApiOkResponse({
+		description: 'If successfully refreshed a new access token',
+		type: RefreshTokenResponse,
+	})
+	@ApiUnauthorizedResponse({
+		description: 'If refresh token is either missing or invalid',
+	})
+	@ApiInternalServerErrorResponse({ description: 'If failed to refresh a new access token' })
 	@Public()
 	@UseGuards(RefreshTokenGuard)
 	@Get(Routes.REFRESH_TOKEN)
@@ -23,7 +42,7 @@ export class RefreshTokenController {
 	public async refreshToken(
 		@CurrentUserId() userId: string,
 		@CurrentUserEmail() userEmail: string,
-	): Promise<IRefreshTokenResponse> {
+	): Promise<RefreshTokenResponse> {
 		this.logger.log(`Will try to generate access token for a user with an Id: "${userId}"`);
 
 		const accessToken = await this.authService.generateJwtToken(userId, userEmail, JwtTokenType.Access);