+++ apps/frontend/src/App.view.tsx
@@ -1,21 +1,24 @@
 import React, { Suspense } from 'react';
 import { Route, BrowserRouter, Routes, Navigate } from 'react-router-dom';
 
-interface Props {
-	isAuthenticated: boolean | null;
+interface IProps {
+	readonly isAuthenticated: boolean | null;
 }
 
 const Auth = React.lazy(() => import('./pages/Auth'));
 const ExternalAuthRedirect = React.lazy(() => import('./pages/ExternalAuthRedirect'));
 const GroupCenter = React.lazy(() => import('./pages/GroupCenter'));
 
+const UserSettings = React.lazy(() => import('./pages/UserSettings'));
 const CliAuth = React.lazy(() => import('./pages/CliAuth'));
 const CliAuthenticated = React.lazy(() => import('./pages/CliAuthenticated'));
 const NotFound = React.lazy(() => import('./pages/NotFound'));
 
-const AppView: React.FC<Props> = (props: React.PropsWithChildren<Props>) => (
+const AppView: React.FC<IProps> = (props: React.PropsWithChildren<IProps>) => (
 	<BrowserRouter>
 		<Suspense fallback={null}>
+			<div id="backdrop-root" />
+			<div id="overlay-root" />
 			<Routes>
 				{props.isAuthenticated === false && (
 					<>
@@ -25,6 +28,7 @@ const AppView: React.FC<Props> = (props: React.PropsWithChildren<Props>) => (
 					</>
 				)}
 				{props.isAuthenticated && <Route path="/group-center/*" element={<GroupCenter />} />}
+				{props.isAuthenticated && <Route path="/user-settings" element={<UserSettings />} />}
 				<Route path="/cli-auth" element={<CliAuth />} />
 				<Route path="/cli-authenticated" element={<CliAuthenticated />} />
 				<Route path="/not-found" element={<NotFound />} />